(rust): Move validation out to a separate function
This looks a bit pointless on its own, but it means we can greatly
simplify the unit tests for the validation logic now we don't need to
call the whole handler.

This will be very useful when we add the functionality to write to
DynamoDB as I'm not going to write tests that invoke the function with a
mocked AWS client. This is a prototype that I'm building in my spare
time between 'real' work. It's not a good use of that limited time to
write loads of tests that mock simple behaviour.