(infra): Set reserved executions and batch size
Reserved executions is now 10 and the batch size is 5.

When I come to do the analysis for this prototype, I expect I'll want to
look at:

1. How cold starts vary across languages
2. How languages with an async pattern compare to those without
3. How the execution time varys per language

In order to do that, I'll need lots of invocations which include a good
number of cold starts. By setting a lower batch size and a reasonably
high concurrency limit I can add more messages to the queue and force
more cold starts